Introduce logging via fast-logger
Unfortunately, CobSession has become a bit more complex in that it has
to know the runtime verbosity to log, and must contain the logger
itself. Furthermore, we had to update the session-creation-functions to
taking a closure within which the Session is available.

The reason for the bracket style CobSession is that the logger is
associated to a finalizer which must be run to ensure buffers are
flushed etc.